+++ src/mono/sample/wasm/browser-bench/Common/Wasm.Bench.Common.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>$(NetCoreAppCurrent)</TargetFramework>
+    <SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+</Project>
