+++ src/mono/sample/wasm/browser-bench/Console/Wasm.Console.Bench.Main.csproj
@@ -9,8 +9,8 @@
   </PropertyGroup>
 
   <PropertyGroup>
-    <_SampleProject>Wasm.Console.Bench.Sample.csproj</_SampleProject>
-    <_SampleAssembly>Wasm.Console.Bench.Sample.dll</_SampleAssembly>
+    <_SampleProject>Wasm.Console.Bench.Main.csproj</_SampleProject>
+    <_SampleAssembly>Wasm.Console.Bench.Main.dll</_SampleAssembly>
     <SignAssembly>False</SignAssembly>
   </PropertyGroup>
 
@@ -20,5 +20,9 @@
     <Compile Include="../*.cs" />
     <Compile Remove="../Browser.cs" />
     <PackageReference Include="Mono.Options" Version="6.12.0.148" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Wasm.Bench.Common.csproj" />
   </ItemGroup>
 </Project>
