+++ src/mono/sample/wasm/browser-bench/Frame/Wasm.Browser.Bench.Frame.csproj
@@ -1,0 +1,7 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>$(NetCoreAppCurrent)</TargetFramework>
+    <SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings>
+    <WasmMainJSPath>frame-main.js</WasmMainJSPath>
+  </PropertyGroup>
+</Project>
