+++ src/mono/sample/wasm/browser-bench/Frame/appstart-frame.html
