+++ src/mono/sample/wasm/browser-bench/JSInterop/Wasm.Console.Bench.JSInterop.csproj
@@ -1,0 +1,10 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>$(NetCoreAppCurrent)</TargetFramework>
+    <SuppressTrimAnalysisWarnings>true</SuppressTrimAnalysisWarnings>
+    <OutputType>Library</OutputType>
+  </PropertyGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Common\Wasm.Console.Bench.Common.csproj" />
+  </ItemGroup>
+</Project>
