+++ src/mono/sample/wasm/browser-bench/Main/Browser.cs
@@ -1,13 +1,13 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Sample
+namespace Wasm.Bench;
+
+public partial class Test
 {
-    public partial class Test
+    public static int Main(string[] args)
     {
-        public static int Main(string[] args)
-        {
-            return 0;
-        }
+        return 0;
     }
 }
+
